+++ packages/api/src/api/routes/users/__init__.py
@@ -1,4 +1,5 @@
 from .auth_router import router as auth_router
+from .roles_router import router as roles_router
 from .users_router import router as users_router
 
-__all__ = ["auth_router", "users_router"]
+__all__ = ["auth_router", "roles_router", "users_router"]